make env production ready

+++ backend/src/index.ts
@@ -7,7 +7,7 @@ import mongoose from 'mongoose';
 import { Errback } from 'express';
 
 import app from './server';
-import { PORT, DATABASE_URL } from './utils/env/env';
+import { PORT, DATABASE_URL, ENVIRONMENT } from './utils/env/env';
 import config from './utils/config/mongodb';
 import getLogger from './utils/logger/logger';
 
@@ -29,6 +29,8 @@ mongoose.connect(DATABASE_URL!, config, (error: any) => {
     logger.error(`Error while connecting to Database: ${JSON.stringify(error)}`);
   } else {
     logger.info('Connected to Database');
-    deleteTestData();
+    if (ENVIRONMENT !== 'prod') {
+      deleteTestData();
+    }
   }
 });
